+++ test/games/prime_hunters/layout/test_prime_hunters_describer.py
@@ -35,7 +35,7 @@ def test_hunters_format_params(octoliths) -> None:
     # Assert
     assert dict(result) == {
         "Logic Settings": ["All tricks disabled"],
-        "Item Pool": [f"Size: {58 + octoliths.placed_octoliths} of 66", "Vanilla starting items"],
+        "Pickup Pool": [f"Size: {58 + octoliths.placed_octoliths} of 66", "Unmodified starting pickup"],
         "Gameplay": [
             "Starts at Celestial Archives - Celestial Gateway",
             "Force Fields: Vanilla",