+++ test/gui/preset_settings/test_pickup_pool_tab.py
@@ -8,13 +8,13 @@ import pytest
 
 from randovania.game_description import default_database
 from randovania.gui.lib.signal_handling import set_combo_with_value
-from randovania.gui.preset_settings.item_pool_tab import PresetItemPool
+from randovania.gui.preset_settings.pickup_pool_tab import PresetPickupPool
 from randovania.interface_common.preset_editor import PresetEditor
 from randovania.layout.base.available_locations import RandomizationMode
 from randovania.layout.base.standard_pickup_state import StandardPickupState
 
 
-def test_on_default_item_updated(skip_qtbot, echoes_game_description, preset_manager):
+def test_on_default_pickup_updated(skip_qtbot, echoes_game_description, preset_manager):
     # Setup
     game = echoes_game_description.game
     base = preset_manager.default_preset_for_game(game).get_preset()
@@ -22,25 +22,25 @@ def test_on_default_item_updated(skip_qtbot, echoes_game_description, preset_man
     options = MagicMock()
 
     editor = PresetEditor(preset, options)
-    window = PresetItemPool(editor, echoes_game_description, MagicMock())
+    window = PresetPickupPool(editor, echoes_game_description, MagicMock())
 
     pickup_database = default_database.pickup_database_for_game(window.game)
-    item = pickup_database.standard_pickups["Dark Beam"]
+    pickup = pickup_database.standard_pickups["Dark Beam"]
     category = pickup_database.pickup_categories["beam"]
     combo = window._default_pickups[category]
 
     # Run
-    set_combo_with_value(combo, item)
+    set_combo_with_value(combo, pickup)
 
     # Assert
-    assert editor.standard_pickup_configuration.default_pickups[category] == item
-    assert editor.standard_pickup_configuration.pickups_state[item] == StandardPickupState(
+    assert editor.standard_pickup_configuration.default_pickups[category] == pickup
+    assert editor.standard_pickup_configuration.pickups_state[pickup] == StandardPickupState(
         num_included_in_starting_pickups=1, included_ammo=(50,)
     )
 
 
 @pytest.mark.parametrize("randomization_mode", RandomizationMode)
-def test_item_pool_count_label(
+def test_pickup_pool_count_label(
     skip_qtbot, blank_game_description, preset_manager, randomization_mode: RandomizationMode
 ):
     base = preset_manager.default_preset_for_game(blank_game_description.game).get_preset()
@@ -61,7 +61,7 @@ def test_item_pool_count_label(
     )
     options = MagicMock()
     editor = PresetEditor(preset, options)
-    window = PresetItemPool(editor, blank_game_description, MagicMock())
+    window = PresetPickupPool(editor, blank_game_description, MagicMock())
 
     # Run
     window.on_preset_changed(preset)
@@ -71,4 +71,4 @@ def test_item_pool_count_label(
     if randomization_mode is RandomizationMode.MAJOR_MINOR_SPLIT:
         message += "<br />Major: 6/6 - Minor: 2/2 - Starting: 1"
 
-    assert window.item_pool_count_label.text() == message
+    assert window.pickup_pool_count_label.text() == message